requirements.txt
pandas
scikit-learn
azure-storage-blob
joblib